+++ CHANGELOG.md
@@ -5,8 +5,11 @@
 - **Web Scraper** Add extraction of repository hosting URL in .nvchecker.toml as an prioritized alternative to the upstream URL
 - **GitLab API** Add method to fetch and decode file content from GitLab API
 - **GitLab API** Refactor error handling
+- **Test** Add tests for getting the closest package tag
 
 ### Bug fixes
+
+- **Package Handler** Improve method to get the closest package tag. This fixes some KDE packages
 
 # 1.0.8 (2025-06-15)